Add Atbash Cipher exercise
This commit introduces a new exercise, the Atbash Cipher, an ancient encryption system.
The exercise consists of implementing encoding and decoding functions in AWK language and includes a testing set written in BATS.
The cipher relies on a simple letter substitution, where each letter is replaced by the equivalent position letter from the end of the alphabet.