Add Automated Readability Index exercise and tests
This commit introduces the Automated Readability Index exercise. It provides students a chance to implement a program in AWK language which will measure the complexity of a given text using the Automated Readability Index formula. In addition, the exercise also includes tests written in BATS to validate the solution's correctness.