Add change calculation feature with tests and README
Introduced a new feature emulating a change-giving scenario in AWK language. The feature includes an AWK script that calculates the minimal number of coins needed to give a certain amount of change. The commit also includes a comprehensive BATS test suite to verify the functionality of the change calculation, as well as a README file detailing the instructions and edge cases.