Add Darts scoring feature with tests and README
Implemented a new feature for Darts game scoring in AWK. This feature includes an AWK script that calculates the score based on the coordinates of dart throws. It also provides a BATS test suite for verifying the functionality of the scoring system, accompanied by a README file that explains the instructions and rules associated with the scoring system.