Add robot simulator code and tests
Added code files and tests for a new exercise, "Robot Simulator". The exercise is about simulating a robot's movement on a virtual grid with AWK programming. The robot can turn right, turn left, and advance. In the solution, handling of different directions and instructions was implemented. The provided tests cover various scenarios to validate the correctness of the solution. Also, a README.md file has been added to provide instructions for the exercise.