Implement grains calculation in AWK
Introduced a new script (grains.awk) to calculate the number of grains on a chessboard, where the number of grains doubles on each square. The script contains bread-and-butter functions to calculate the number of grains on a given square and the total number of grains on the board. Part of the implementation includes error checking to ensure that the input square is within the range of 1 to 64.